+++ app/src/views/articles/form.js
@@ -1,7 +1,7 @@
 import Marionette from 'backbone.marionette';
 import Article from 'models/article';
 import FormBehavior from 'behaviors/form';
-import Session from 'facades/session';
+import Session from 'services/session';
 import user from 'helpers/user';
 import template from 'templates/articles/form';
 import { props } from 'decorators';